scrum-assessment-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
ecdsa==0.19.1
fastapi==0.115.12
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
idna==3.10
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
